# More efficient serialization of documents with ASCII keys