Before:
```
[info] Benchmark                      (size)   Mode  Cnt         Score        Error  Units
[info] GoogleMapsAPIWriting.smithy4s     N/A  thrpt    5     39816.307 ±    983.413  ops/s
[info] NestedStructsWriting.smithy4s     128  thrpt    5    403630.790 ±  10781.265  ops/s
[info] PrimitivesWriting.smithy4s        N/A  thrpt    5   5866595.809 ± 208961.868  ops/s
```
After:
```
[info] Benchmark                      (size)   Mode  Cnt         Score        Error  Units
[info] GoogleMapsAPIWriting.smithy4s     N/A  thrpt    5     46183.832 ±    877.871  ops/s
[info] NestedStructsWriting.smithy4s     128  thrpt    5    460022.818 ±  41012.061  ops/s
[info] PrimitivesWriting.smithy4s        N/A  thrpt    5   6737206.295 ± 516062.120  ops/s
```